requirements.txt
django==3.1.11
djangorestframework==3.12.4
coreapi==2.3.3
PyYAML==5.4.1